backend/requirements.txt
django>=5.0
django-ninja
djangorestframework-simplejwt
